# perf(client): override PagingList Mixin methods from "MutableSequence"