Results:
| method         | before           | after             |
| -------------- | ---------------- | ----------------- |
| `__iter__`     | 629 ns ± 21.5 ns | 279 ns ± 30.6 ns  |
| `__reversed__` | 638 ns ± 25.2 ns | 256 ns ± 15.3 ns  |
| `__contains__` | 5.81 ms ± 108 µs | 1.55 ms ± 8.22 µs |
| `__iadd__`     | 4.02 µs ± 576 ns | 1.78 µs ± 89.1 ns |
| `index`        | 5.7 ms ± 73 µs   | 1.99 ms ± 47.1 µs |
| `count`        | 5.82 s ± 7.8 ms  | 1.59 s ± 14.7 ms  |
| `reverse`      | 12.8 s ± 28.6 ms | 5.3 ms ± 186 µs   |
| `append`       | 1.12 µs ± 31 ns  | 735 ns ± 94.7 ns  |
| `extend`       | 4.1 µs ± 659 ns  | 2.09 µs ± 234 ns  |
| `pop`          | 1.33 µs ± 150 ns | 600 ns ± 43.1 ns  |